feat: update index for VIP-499
